+++ src/window.rs
@@ -597,6 +597,16 @@ impl Window {
         self.window.set_maximized(maximized)
     }
 
+    /// Gets the window's current maximized state.
+    ///
+    /// ## Platform-specific
+    ///
+    /// - **iOS / Android / Web:** Unsupported.
+    #[inline]
+    pub fn maximized(&self) -> bool {
+        self.window.maximized()
+    }
+
     /// Sets the window to fullscreen or back.
     ///
     /// ## Platform-specific